Add BookingController with GetAllBookings endpoint
The `<ItemGroup>` defining the `Controllers\` folder was removed
from the `Camping-BookingSystem.csproj` file, simplifying the
project structure.

A new `BookingController` class was introduced in the
`Camping_BookingSystem.Controllers` namespace. It is an API
controller with `[ApiController]` and `[Route]` attributes,
designed to handle HTTP requests.

The controller depends on the `IBookingRepository` interface,
injected via the constructor, adhering to the Dependency
Injection pattern.

A `GetAllBookings` asynchronous HTTP GET method was added to
retrieve all bookings and return them with an `Ok` status.

Necessary `using` directives were added to support the new
functionality.